pyproject.toml
[tool.poetry]
name = "fox"
version = "0.1.0"
description = "Normalizing Flows in JAX"
authors = ["Marin Vlastelica <marin.vlastelicap@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
jax = "*"
optax = "*"
tqdm = "*"
flax = "*"
sklearn = "*"
numpyro = "*"
dm-haiku = "^0.0.5"
setuptools = "*"
pytest = "*"
fox = {path = ".", develop = true}

[tool.poetry.dev-dependencies]

matplotlib = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
